add balanceOf util function
