+++ src/quote/provider/icusd.ts
@@ -1,4 +1,5 @@
 import { BigNumber } from '@ethersproject/bignumber'
+import { Address } from 'viem'
 
 import { TheUSDCYieldIndex } from 'constants/tokens'
 import {
@@ -6,9 +7,8 @@ import {
   FlashMintWrappedBuildRequest,
   WrappedTransactionBuilder,
 } from 'flashmint'
-import { createClient } from 'utils/clients'
 import { getExpectedReserveRedeemQuantity } from 'utils/custom-oracle-nav-issuance-module'
-import { Address, parseAbi } from 'viem'
+import { getUsdcBalance } from 'utils/erc20'
 
 import { FlashMintNavQuoteProvider } from '../flashmint/nav'
 import { WrappedQuoteProvider } from '../flashmint/wrapped'
@@ -51,7 +51,11 @@ export class IcUsdQuoteRouter
         '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
         indexTokenAmount.toBigInt()
       )
-      const usdcBalance = await this.getUsdcBalance(chainId)
+      const usdcBalance = await getUsdcBalance(
+        '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+        this.icUsd.address as Address,
+        chainId
+      )
       console.log(usdcBalance.toString(), 'USDC')
       // 80% of the USDC balance of icUSD
       const threshold = (usdcBalance * 80n) / 100n
@@ -71,20 +75,6 @@ export class IcUsdQuoteRouter
       if (useFlashMintNav) return await this.getFlashMintNavQuote(request)
       return await this.getFlashMintWrappedQuote(request)
     }
-  }
-
-  private async getUsdcBalance(chainId: number) {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const publicClient = createClient(chainId)!
-    const amount: bigint = (await publicClient.readContract({
-      address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
-      abi: parseAbi([
-        'function balanceOf(address account) view returns (uint256)',
-      ]),
-      functionName: 'balanceOf',
-      args: ['0x54EE8A49155F701F0d5Ff088CD36fbBF1a5B9f44'],
-    })) as bigint
-    return amount
   }
 
   private async getFlashMintNavQuote(request: IcUsdQuoteRequest) {
