+++ src/utils/erc20.ts
@@ -1,0 +1,20 @@
+import { Address, parseAbi } from 'viem'
+import { createClient } from './clients'
+
+export async function getUsdcBalance(
+  token: Address,
+  account: Address,
+  chainId: number
+) {
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const publicClient = createClient(chainId)!
+  const amount: bigint = (await publicClient.readContract({
+    address: token,
+    abi: parseAbi([
+      'function balanceOf(address account) view returns (uint256)',
+    ]),
+    functionName: 'balanceOf',
+    args: [account],
+  })) as bigint
+  return amount
+}
